requirements.txt
pygame
pyyaml
pyOpengl